+++ lib/src/speech_to_text/speech_sub_bar.dart
@@ -1,23 +1,58 @@
+// Copyright (c) 2022, Alexqwesa.
+// All rights reserved. Use of this source code
+// is governed by a BSD-style license that can be found in the LICENSE file.
+
+import 'package:app_bar_with_search_switch/src/app_bar_with_search_switch.dart';
 import 'package:flutter/material.dart';
 import 'package:speech_to_text/speech_to_text.dart';
 
-/// Indicator(controller) shown when speech recognition is active.
+/// Indicator(and controller) shown when speech recognition is active.
+///
+/// It supposed to be shown when [SpeechToText] is already in process of activating.
+/// If you want to use you own SubBar for [AppBarWithSearchSwitch],
+/// assign it to parameter [AppBarWithSearchSwitch].[speechSubBar] as in
+/// template bellow:
+///```dart
+///  // inside AppBarWithSearchSwitch use:
+///  // speechSubBarHeight: kToolbarHeight * 1.2,
+///  speechSubBar: (context) {
+///    final speech = AppBarWithSearchSwitch.of(context)!.speechEngine;
+///    final isListening = AppBarWithSearchSwitch.of(context)!.isListeningToSpeech;
+///    final textNotifier = AppBarWithSearchSwitch.of(context)!.textNotifier;
+///
+///    return ValueListenableBuilder(
+///      valueListenable: isListening,
+///      builder: (context, _, child) {
+///        return FloatingActionButton(
+///          backgroundColor: speech.isListening ? Colors.red : Colors.grey,
+///          autofocus: true,
+///          onPressed: () async {
+///            if (speech.isListening) {
+///              await speech.stop();
+///              isListening.value = false; // not necessary, just failsafe
+///            } else {
+///              await innerStartListening(
+///                speech: speech,
+///                isListening: isListening,
+///                textNotifier: textNotifier,
+///              );
+///            }
+///          },
+///          child: const Icon(Icons.mic_rounded),
+///        );
+///      },
+///    );
+///  },
+///```
 class SpeechSubBar extends StatelessWidget {
-  final ValueNotifier<String> textNotifier;
-  final Future<void> Function() speechStartListening;
-  final ValueNotifier<bool> isListen;
-  final SpeechToText speech;
-
-  const SpeechSubBar({
-    required this.textNotifier,
-    required this.speechStartListening,
-    required this.isListen,
-    required this.speech,
-    Key? key,
-  }) : super(key: key);
+  const SpeechSubBar({Key? key}) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
+    final speech = AppBarWithSearchSwitch.of(context)!.speechEngine;
+    final isListening = AppBarWithSearchSwitch.of(context)!.isListeningToSpeech;
+    final textNotifier = AppBarWithSearchSwitch.of(context)!.textNotifier;
+
     if (!(speech.isListening || speech.isAvailable)) {
       return const Center(
         child: Text('Speech recognition is initializing...'),
@@ -31,7 +66,7 @@ class SpeechSubBar extends StatelessWidget {
         child: Column(
           children: [
             ValueListenableBuilder(
-              valueListenable: isListen,
+              valueListenable: isListening,
               builder: (context, _, child) {
                 return FloatingActionButton(
                   backgroundColor:
@@ -40,9 +75,13 @@ class SpeechSubBar extends StatelessWidget {
                   onPressed: () async {
                     if (speech.isListening) {
                       await speech.stop();
-                      isListen.value = false;
+                      isListening.value = false; // not necessary, just failsafe
                     } else {
-                      await speechStartListening();
+                      await innerStartListening(
+                        speech: speech,
+                        isListening: isListening,
+                        textNotifier: textNotifier,
+                      );
                     }
                   },
                   child: const Icon(Icons.mic_rounded),
@@ -58,4 +97,22 @@ class SpeechSubBar extends StatelessWidget {
       ),
     );
   }
+}
+
+/// This function should only be used inside [AppBarWithSearchSwitch].[speechSubBar],
+///
+/// it will call [SpeechToText].[listen] inside...
+Future<void> innerStartListening(
+    {required SpeechToText speech,
+    required ValueNotifier<bool> isListening,
+    required ValueNotifier<String> textNotifier}) async {
+  isListening.value = speech.isListening;
+  final previousText = textNotifier.value; // don't duplicate results, but save current field
+  await speech.listen(
+    onResult: (result) {
+      textNotifier.value = previousText + result.recognizedWords;
+      isListening.value = speech.isListening;
+    },
+  );
+  isListening.value = speech.isListening;
 }